+++ PDFFileMetadataReader/metadata.csv
@@ -1,8 +1,5 @@
 Name,Author,File Size (bytes),Page Size,Pages,Title,Subject,Keywords,FileType,Version,Source application,Images,Fonts
-Circuitos Lab.1.pdf, Elisa Ajxup Yax, 428920, Carta, 3, Sin título, Sin asunto, Sin palabras clave, PDF, 1.7, Microsoft® Word for Microsoft 365, 7, 0, 
-468_2004_ESCA-ST_MAESTRIA_Alvarado__Maldonado.pdf, Guadalupe Alvarado, 6024302, Carta, 287, RESUMEN, Sin asunto, Sin palabras clave, PDF, 1.4, Acrobat PDFMaker 5.0 para Word, 105, 70, 
-Arboles B-3.pdf, Leonel Enriquez, 3917170, 20 x 11.2, 35, Arboles B, Sin asunto, DAFJl7cFYWA,BADQseUAyT0, PDF, 1.4, Canva, 21, 0, 
-san_marcos.pdf, Margarita Castillo, 520192, Carta, 17, Departamento de Chiquimula, Sin asunto, Sin palabras clave, PDF, 1.5, Microsoft® Word 2013, 18, 9, 
-Myipymedeguatemala2015-2017.pdf, Angel, 6495184, Carta, 24, Sin título, Sin asunto, Sin palabras clave, PDF, 1.7, Adobe InDesign CC 14.0 (Macintosh), 150, 14, 
-sistema_nacional_de_informacion_mipyme_guatemala_ano_base_2015.pdf, Sin autor, 1334952, Carta, 32, Sin título, Sin asunto, Sin palabras clave, PDF, 1.7, Adobe InDesign CC 2017 (Macintosh), 38, 27, 
+Myipymedeguatemala2015-2017.pdf, Sin autor, 6495184, Carta, 24, Sin título, Sin asunto, Sin palabras clave, PDF, 1.7, Adobe InDesign CC 14.0 (Macintosh), 150, 14, 
+san_marcos.pdf, Margarita Castillo, 520192, Carta, 17, Departamento de Chiquimula, Investigación, Sin palabras clave, PDF, 1.5, Microsoft® Word 2013, 18, 9, 
+sistema_nacional_de_informacion_mipyme_guatemala_ano_base_2015.pdf, Karol Castro, 1334952, Carta, 32, Sin título, Sin asunto, Sin palabras clave, PDF, 1.7, Adobe InDesign CC 2017 (Macintosh), 38, 27, 
 Tarea 2 - Elisa Yanira Ajxup Yax.pdf, CamScanner, 1086108, Oficio, 5, Tarea 2 - Elisa Yanira Ajxup Yax, Tarea 2 - Elisa Yanira Ajxup Yax, Sin palabras clave, PDF, 1.7, Sin aplicación de origen, 10, 0,
